# Alternate height encoding