--- a/addons/zylann.hterrain/hterrain_data.gd
+++ b/addons/zylann.hterrain/hterrain_data.gd
@@ -1700,7 +1700,7 @@ static func get_map_shader_param_name(map_type: int, index: int) -> String:
 #	return null
 
 
-const _V2_UNIT_STEPS = 100.0
+const _V2_UNIT_STEPS = 1024.0
 const _V2_MIN = -8192.0
 
 
@@ -1722,6 +1722,20 @@ static func encode_height_to_rgb8_unorm(h: float) -> Color:
 	var b := float((hi>>16)&0xff) / 255.0
 
 	return Color(r, g, b, 1.0)
+
+const _V2_MAX = 8191.0
+const _V2_DF = 255.0 / _V2_UNIT_STEPS
+
+static func prev_decode_height_from_rgb8_unorm(c: Color) -> float:
+	return (c.r * 0.25 + c.g * 64.0 + c.b * 16384.0) * (4.0 * _V2_DF) + _V2_MIN
+
+static func prev_encode_height_to_rgb8_unorm(h: float) -> Color:
+	h -= _V2_MIN
+	var i := int(h * _V2_UNIT_STEPS)
+	var r := i % 256
+	var g := (i / 256) % 256
+	var b := i / 65536
+	return Color(r, g, b, 255.0) / 255.0
 
 
 static func convert_heightmap_to_float(src: Image, logger: HT_Logger.HT_LoggerBase) -> Image: