--- a/addons/zylann.hterrain/tools/importer/importer_dialog.gd
+++ b/addons/zylann.hterrain/tools/importer/importer_dialog.gd
@@ -16,8 +16,11 @@ signal permanent_change_performed(message)
 @onready var _warnings_label : Label = \
 	$VBoxContainer/ColorRect/ScrollContainer/VBoxContainer/Warnings
 
-const RAW_LITTLE_ENDIAN = 0
-const RAW_BIG_ENDIAN = 1
+
+enum {
+	RAW_LITTLE_ENDIAN,
+	RAW_BIG_ENDIAN
+}
 
 var _terrain : HTerrain = null
 var _logger = HT_Logger.get_for(self)
@@ -37,7 +40,13 @@ func _ready():
 		"raw_endianess": {
 			"type": TYPE_INT,
 			"usage": "enum",
-			"enum_items": ["Little Endian", "Big Endian"],
+			"enum_items": [[RAW_LITTLE_ENDIAN, "Little Endian"], [RAW_BIG_ENDIAN, "Big Endian"]],
+			"enabled": false
+		},
+		"bit_depth": {
+			"type": TYPE_INT,
+			"usage": "enum",
+			"enum_items": [[HTerrainData.BIT_DEPTH_UNDEFINED, "Undefined"], [HTerrainData.BIT_DEPTH_16, "16-bit"], [HTerrainData.BIT_DEPTH_24, "24-bit"]],
 			"enabled": false
 		},
 		"min_height": {
@@ -61,7 +70,8 @@ func _ready():
 			"exts": ["png"]
 		}
 	})
-	
+	_inspector.set_item_disabled("bit_depth", HTerrainData.BIT_DEPTH_UNDEFINED, true)
+
 	# Testing
 #	_errors_label.text = "- Hello World!"
 #	_warnings_label.text = "- Yolo Jesus!"
@@ -137,7 +147,8 @@ func _on_ImportButton_pressed():
 			"path": heightmap_path,
 			"min_height": _inspector.get_value("min_height"),
 			"max_height": _inspector.get_value("max_height"),
-			"big_endian": endianess == RAW_BIG_ENDIAN
+			"big_endian": endianess == RAW_BIG_ENDIAN,
+			"bit_depth": _inspector.get_value("bit_depth"),
 		}
 
 	var colormap_path = _inspector.get_value("colormap")
@@ -168,6 +179,34 @@ func _on_Inspector_property_changed(key: String, value):
 	if key == "heightmap":
 		var is_raw = value.get_extension().to_lower() == "raw"
 		_inspector.set_property_enabled("raw_endianess", is_raw)
+		_inspector.set_property_enabled("bit_depth", is_raw)
+		var reset_bit_depth:bool = true
+		if is_raw:
+			var bit_depth:int = _estimate_bit_depth_for_raw_file(value)
+			if bit_depth >= HTerrainData.BIT_DEPTH_UNDEFINED:
+				_inspector.set_value("bit_depth", bit_depth)
+				reset_bit_depth = false
+		if reset_bit_depth:
+			_inspector.set_value("bit_depth", HTerrainData.BIT_DEPTH_UNDEFINED)
+
+
+# _estimate_bit_depth_for_raw_file returns the file's identified bit depth, or 0.
+static func _estimate_bit_depth_for_raw_file(path: String) -> int:
+	var ext := path.get_extension().to_lower()
+	if ext == "raw":
+		var f := FileAccess.open(path, FileAccess.READ)
+		if f == null:
+			return HTerrainData.BIT_DEPTH_UNDEFINED
+
+		var file_len := f.get_length()
+		f = null # close file
+
+		for bit_depth in [HTerrainData.BIT_DEPTH_16, HTerrainData.BIT_DEPTH_24]:
+			var file_res := HT_Util.integer_square_root(file_len / (bit_depth/8))
+			if file_res > 0:
+				return bit_depth
+
+	return HTerrainData.BIT_DEPTH_UNDEFINED
 
 
 func _validate_form() -> HT_ErrorCheckReport:
@@ -176,6 +215,7 @@ func _validate_form() -> HT_ErrorCheckReport:
 	var heightmap_path : String = _inspector.get_value("heightmap")
 	var splatmap_path : String = _inspector.get_value("splatmap")
 	var colormap_path : String = _inspector.get_value("colormap")
+	var bit_depth = _inspector.get_value("bit_depth")
 
 	if colormap_path == "" and heightmap_path == "" and splatmap_path == "":
 		res.errors.append("No maps specified.")
@@ -195,7 +235,7 @@ func _validate_form() -> HT_ErrorCheckReport:
 			# so we avoid loading other maps everytime to do further checks
 			return res
 
-		var image_size_result = _load_image_size(heightmap_path, _logger)
+		var image_size_result = _load_image_size(heightmap_path, _logger, bit_depth)
 		if image_size_result.error_code != OK:
 			res.errors.append(str("Cannot open heightmap file: ", image_size_result.to_string()))
 			return res
@@ -211,18 +251,18 @@ func _validate_form() -> HT_ErrorCheckReport:
 		heightmap_size = adjusted_size
 
 	if splatmap_path != "":
-		_check_map_size(splatmap_path, "splatmap", heightmap_size, res, _logger)
+		_check_map_size(splatmap_path, "splatmap", heightmap_size, bit_depth, res, _logger)
 
 	if colormap_path != "":
-		_check_map_size(colormap_path, "colormap", heightmap_size, res, _logger)
+		_check_map_size(colormap_path, "colormap", heightmap_size, bit_depth, res, _logger)
 
 	return res
 
 
-static func _check_map_size(path: String, map_name: String, heightmap_size: int, 
+static func _check_map_size(path: String, map_name: String, heightmap_size: int, bit_depth: int, 
 	res: HT_ErrorCheckReport, logger):
 	
-	var size_result := _load_image_size(path, logger)
+	var size_result := _load_image_size(path, logger, bit_depth)
 	if size_result.error_code != OK:
 		res.errors.append(str("Cannot open splatmap file: ", size_result.to_string()))
 		return
@@ -251,7 +291,7 @@ class HT_ImageSizeResult:
 		return HT_Errors.get_message(error_code)
 
 
-static func _load_image_size(path: String, logger) -> HT_ImageSizeResult:
+static func _load_image_size(path: String, logger, bit_depth: int) -> HT_ImageSizeResult:
 	var ext := path.get_extension().to_lower()
 	var result := HT_ImageSizeResult.new()
 
@@ -269,6 +309,11 @@ static func _load_image_size(path: String, logger) -> HT_ImageSizeResult:
 		return result
 
 	elif ext == "raw":
+		if bit_depth != HTerrainData.BIT_DEPTH_16 && bit_depth != HTerrainData.BIT_DEPTH_24:
+			result.error_code = ERR_INVALID_DATA
+			result.error_message = "Invalid bit depth for .raw file import: {0}. Only 16-bit or 24-bit supported...".format([bit_depth])
+			return result
+
 		var f := FileAccess.open(path, FileAccess.READ)
 		if f == null:
 			var err := FileAccess.get_open_error()
@@ -276,14 +321,13 @@ static func _load_image_size(path: String, logger) -> HT_ImageSizeResult:
 			result.error_code = err
 			return result
 
-		# Assume the raw data is square in 16-bit format,
-		# so its size is function of file length
+		# Assume the raw data is square, so its size is function of file length
 		var flen := f.get_length()
 		f = null
-		var size_px = HT_Util.integer_square_root(flen / 2)
+		var size_px = HT_Util.integer_square_root(flen / (bit_depth/8))
 		if size_px == -1:
 			result.error_code = ERR_INVALID_DATA
-			result.error_message = "RAW image is not square"
+			result.error_message = "RAW image is not square or your bit depth choice is incorrect…"
 			return result
 		
 		logger.debug("Deduced RAW heightmap resolution: {0}*{1}, for a length of {2}" \