Fixes https://github.com/Zylann/godot_heightmap_plugin/issues/368
* still WIP -> precision more or less the same
* new algorithm is simpler and faster and output seems more consistent
* breaks compatibility with existing heightmaps -> need upgrade func

**NOTE**: based on my PR https://github.com/Zylann/godot_heightmap_plugin/pull/369 to allow me to import heightmaps of sufficient quality. Will rebase once that PR has been merged  
Therefore, please only see newest commit: _Change heightmap encoding_

---

On import:
* min height: 0.0
* max height: 131.87

On Export, calculated:
* min height: 0.0
* max height: 131.78027

before, auto-calculated heights on export were:
* min height: 0.0004844
* max height: 131.78074


## Benchmarking

new encoding+decoding: max diff 0-0.000977 over 7229.061 msecs
previous encoding+decoding: max diff 0-0.`000975` over 7304.282 msecs
new encoding+decoding: max diff 0-0.000977 over 7148.446 msecs
previous encoding+decoding: max diff 0-0.`000976` over 7224.242 msecs
new encoding+decoding: max diff 0-0.000977 over 7099.373 msecs
previous encoding+decoding: max diff 0-0.000975 over 7174.645 msecs

**current progress result: Basically equal precision loss but faster and output seems consistent** (rounding in current logic seems to round differently sometimes)