@Zylann 2023-06-10
FYI https://github.com/Zylann/godot_heightmap_plugin/issues/316#issuecomment-1585174443