@MJacred 2023-06-16
will close this PR due to 32-bit support (https://github.com/Zylann/godot_heightmap_plugin/commit/e49ac9a8e6a9cf41667c6722d663cd6370026df5)